+++ src/main/java/ee/sk/smartid/SignatureResponseValidator.java
@@ -49,6 +49,11 @@ import ee.sk.smartid.rest.dao.SessionResult;
 import ee.sk.smartid.rest.dao.SessionSignature;
 import ee.sk.smartid.rest.dao.SessionSignatureAlgorithmParameters;
 import ee.sk.smartid.rest.dao.SessionStatus;
+import ee.sk.smartid.signature.HashAlgorithm;
+import ee.sk.smartid.signature.MaskGenAlgorithm;
+import ee.sk.smartid.signature.RsaSsaPssParameters;
+import ee.sk.smartid.signature.SigningSignatureAlgorithm;
+import ee.sk.smartid.signature.TrailerField;
 import ee.sk.smartid.util.StringUtil;
 
 /**