+++ src/main/java/ee/sk/smartid/signature/SignableData.java
@@ -1,10 +1,10 @@
-package ee.sk.smartid;
+package ee.sk.smartid.signature;
 
 /*-
  * #%L
  * Smart ID sample Java client
  * %%
- * Copyright (C) 2018 - 2025 SK ID Solutions AS
+ * Copyright (C) 2018 - 2026 SK ID Solutions AS
  * %%
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -12,10 +12,10 @@ package ee.sk.smartid;
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * 
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- *
+ * 
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -32,56 +32,62 @@ import java.util.Base64;
 import ee.sk.smartid.exception.permanent.SmartIdRequestSetupException;
 
 /**
- * This class can be used to contain the hash
- * to be signed
+ * This class can be used to contain the data
+ * to be signed when it is not yet in hashed format
  * <p>
- * {@link SignableData} can be used
- * instead when the data to be signed is not already
+ * {@link SignableHash} can be used
+ * instead when the data to be signed is already
  * in hashed format.
  */
-public record SignableHash(byte[] hashToBeSigned, HashAlgorithm hashAlgorithm) implements Serializable, DigestInput {
+public record SignableData(byte[] dataToSign, HashAlgorithm hashAlgorithm) implements Serializable, DigestInput {
 
     /**
-     * Creates {@link SignableHash} instance,
+     * Creates a new instance of SignableData
      * <p>
      * Will use SHA-512 as the default hashing algorithm
      *
-     * @param hashToSign byte array of hash to be signed
-     * @throws SmartIdRequestSetupException when hashToSign is missing or empty
+     * @param dataToSign byte array of data to be signed
      */
-    public SignableHash(byte[] hashToSign) {
-        this(hashToSign, HashAlgorithm.SHA_512);
+    public SignableData(byte[] dataToSign) {
+        this(dataToSign, HashAlgorithm.SHA_512);
     }
 
     /**
-     * Creates {@link SignableHash} instance
+     * Creates a new instance of SignableData
      *
-     * @param hashToBeSigned byte array of hash to be signed
-     * @param hashAlgorithm  hashing algorithm used to create the hash
-     * @throws SmartIdRequestSetupException when input parameters are missing or empty
+     * @param dataToSign    byte array of data to be signed
+     * @param hashAlgorithm hashing algorithm to be used
+     * @throws SmartIdRequestSetupException when input values are missing or empty
      */
-    public SignableHash(byte[] hashToBeSigned, HashAlgorithm hashAlgorithm) {
-        validateInputs(hashToBeSigned, hashAlgorithm);
-        this.hashToBeSigned = hashToBeSigned.clone();
-        this.hashAlgorithm = hashAlgorithm;
-    }
-
-    private static void validateInputs(byte[] hash, HashAlgorithm hashAlgorithm) {
-        if (hash == null || hash.length == 0) {
-            throw new SmartIdRequestSetupException("Parameter 'hash' cannot be empty");
+    public SignableData(byte[] dataToSign, HashAlgorithm hashAlgorithm) {
+        if (dataToSign == null || dataToSign.length == 0) {
+            throw new SmartIdRequestSetupException("Parameter 'dataToSign' cannot be empty");
         }
         if (hashAlgorithm == null) {
             throw new SmartIdRequestSetupException("Parameter 'hashAlgorithm' must be set");
         }
+        this.dataToSign = dataToSign.clone();
+        this.hashAlgorithm = hashAlgorithm;
     }
 
     /**
-     * Get the hash as Base64-encoded string
+     * Calculates the digest of the data to be signed
+     * and returns it in Base64 encoded format
      *
-     * @return String
+     * @return Base64 encoded hash
      */
     @Override
     public String getDigestInBase64() {
-        return Base64.getEncoder().encodeToString(hashToBeSigned);
+        byte[] digest = calculateHash();
+        return Base64.getEncoder().encodeToString(digest);
+    }
+
+    /**
+     * Calculates the digest of the data to be signed
+     *
+     * @return hash
+     */
+    public byte[] calculateHash() {
+        return DigestCalculator.calculateDigest(dataToSign, hashAlgorithm);
     }
 }