+++ src/test/java/ee/sk/smartid/NotificationAuthenticationSessionRequestBuilderTest.java
@@ -62,6 +62,8 @@ import ee.sk.smartid.rest.SmartIdConnector;
 import ee.sk.smartid.rest.dao.NotificationAuthenticationSessionRequest;
 import ee.sk.smartid.rest.dao.NotificationAuthenticationSessionResponse;
 import ee.sk.smartid.rest.dao.SemanticsIdentifier;
+import ee.sk.smartid.signature.AuthenticationSignatureAlgorithm;
+import ee.sk.smartid.signature.HashAlgorithm;
 
 class NotificationAuthenticationSessionRequestBuilderTest {