+++ src/test/java/ee/sk/smartid/SignatureResponseValidatorTest.java
@@ -51,6 +51,7 @@ import ee.sk.smartid.rest.dao.SessionResultDetails;
 import ee.sk.smartid.rest.dao.SessionSignature;
 import ee.sk.smartid.rest.dao.SessionSignatureAlgorithmParameters;
 import ee.sk.smartid.rest.dao.SessionStatus;
+import ee.sk.smartid.signature.SigningSignatureAlgorithm;
 
 class SignatureResponseValidatorTest {